Relocate files to not need 'require' to import mock file
