+++ src/__mocks__/getLC.js
@@ -2,6 +2,7 @@ const setTasksLocalStorage = (task) => {
   const mockTasksContainer = [];
   mockTasksContainer.push(...task);
   const result = mockTasksContainer;
+  console.log('helllo');
   return result;
 };
